Mlve seeds to header
Makes more sense here, especially sknce seeds are no longer
contained by keys
